requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
pycparser==2.21
PyJWT==2.5.0
python-dotenv==0.21.0
python-telegram-bot==13.14
python-telegram-bot-pagination==0.0.2
python3-openid==3.2.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tornado==6.1
tzdata==2022.2
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
